+++ tests/test_qc_filter.py
@@ -159,7 +159,7 @@ def test_filter_genes(adata):
 
 def test_filter_cells(adata):
     """ Test whether cells were filtered out based on a boolean column"""
-
+    adata = adata.copy()  # copy adata to avoid inplace changes
     adata.obs["cell_bool"] = np.random.choice(a=[False, True], size=adata.shape[0])
     n_false = sum(~adata.obs["cell_bool"])
     qc.filter_cells(adata, "cell_bool", inplace=True)  # removes all genes with boolean True
@@ -167,17 +167,57 @@ def test_filter_cells(adata):
     assert adata.shape[0] == n_false
 
 
-def test_predict_sex(caplog, adata):
-
+@pytest.mark.parametrize("which, to_filter", [("obs", ["AAACCCACAGCCTATA", "AAACCCACAGGGCTTC"]),
+                                              ("var", ["ENSMUSG00000051951", "ENSMUSG00000102851"])])
+def test_filter_object(adata, which, to_filter):
+    """Test whether cells/genes are filtered based on a list of cells/genes."""
     adata = adata.copy()  # copy adata to avoid inplace changes
+    qc._filter_object(adata, to_filter, which=which)
+    table = adata.obs if which == "obs" else adata.var
+    assert all([i not in table.index for i in to_filter])
+
+
+def test_filter_object_fail(adata):
+    """Test whether invalid input raises the correct errors."""
+    adata = adata.copy()
+    adata.obs["notbool"] = np.random.choice(a=[False, True, np.nan], size=adata.shape[0])
+    with pytest.raises(ValueError, match="Column notbool contains values that are not of type boolean"):
+        qc._filter_object(adata, "notbool", which="obs")
+
+    with pytest.raises(ValueError, match="Column invalid not found"):
+        qc._filter_object(adata, "invalid", which="obs")
+
+
+@pytest.mark.parametrize("threshold", [0.3, 0.0])
+def test_predict_sex(caplog, adata, threshold):
+    """Test if predict_sex warns on invalid gene and succeeds."""
+    adata = adata.copy()  # copy adata to avoid inplace changes
+
     # gene not in data
     with caplog.at_level(logging.INFO):
         qc.predict_sex(adata, groupby='sample')
         assert "Selected gene is not present in the data. Prediction is skipped." in caplog.records[1].message
 
     # gene in data
-    qc.predict_sex(adata, gene='Xkr4', gene_column='gene', groupby='sample')
+    qc.predict_sex(adata, gene='Xkr4', gene_column='gene', groupby='sample', threshold=threshold)
     assert 'predicted_sex' in adata.obs.columns
+
+
+def test_predict_sex_diff_types(caplog, adata):
+    """Test predict_sex for different adata.X types."""
+
+    adata_ndarray = adata.copy()
+    adata_ndarray.X = adata_ndarray.X.toarray()
+    adata_matrix = adata_ndarray.copy()
+    adata_matrix.X = np.asmatrix(adata_matrix.X)
+
+    # ndarray
+    qc.predict_sex(adata_ndarray, gene='Xkr4', gene_column='gene', groupby='sample')
+    assert 'predicted_sex' in adata_ndarray.obs.columns
+
+    # matrix
+    qc.predict_sex(adata_matrix, gene='Xkr4', gene_column='gene', groupby='sample')
+    assert 'predicted_sex' in adata_matrix.obs.columns
 
 
 @pytest.mark.parametrize(